Move PCX code to utils/pcx.cpp
