+++ Source/CMakeLists.txt
@@ -107,6 +107,7 @@ set(libdevilutionx_SRCS
   utils/language.cpp
   utils/logged_fstream.cpp
   utils/paths.cpp
+  utils/pcx.cpp
   utils/sdl_bilinear_scale.cpp
   utils/sdl_thread.cpp
   utils/utf8.cpp
