+++ Source/utils/pcx.hpp
@@ -1,6 +1,9 @@
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
+
+#include <SDL.h>
 
 namespace devilution {
 
@@ -24,5 +27,11 @@ struct PCXHeader {
 	uint16_t VscreenSize;
 	uint8_t Filler[54];
 };
+
+static constexpr size_t PcxHeaderSize = 128;
+
+bool LoadPcxMeta(SDL_RWops *handle, int &width, int &height, uint8_t &bpp);
+bool LoadPcxPixelsAndPalette(SDL_RWops *handle, int width, int height, std::uint8_t bpp,
+    uint8_t *buffer, std::ptrdiff_t bufferPitch, SDL_Color *palette);
 
 } // namespace devilution
